+++ Diabetes.Test/Diabetes.Test.csproj
@@ -60,7 +60,7 @@
         </Reference>
     </ItemGroup>
     <ItemGroup>
-        <Compile Include="CalculateCorrectionBolusTests.cs" />
+        <Compile Include="BolusCalculator\CalculateCorrectionBolusTests.cs" />
         <Compile Include="HelperFunctions\DateTimeHelperFunctionsTests\GetSleepingPeriodStartAndEndDateTimesTests.cs" />
         <Compile Include="Properties\AssemblyInfo.cs"/>
     </ItemGroup>
